requirement.txt
numpy>=1.24.0
pypdf>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# 実際の埋め込み生成には以下を追加できます：
# sentence-transformers>=2.2.0
# transformers>=4.30.0
# torch>=2.0.0

# ベクトルストアとして以下も使用可能：
# faiss-cpu>=1.7.4
# chromadb>=0.4.0
# pinecone-client>=2.2.0

# その他のドキュメント読み込み用：
# python-docx>=0.8.11
# beautifulsoup4>=4.12.0

# 検索・ランキング用：
# rank-bm25>=0.2.2
# scikit-learn>=1.3.0

